+++ RELEASE_NOTES.md
@@ -15,5 +15,12 @@ Removed Deprecated Modules
 Deprecations
 ============
 
+* `ocean.io.serialize.StructSerializer`
+
+  The binary serialization facilities of this module have been deprecated. Any
+  code which is still using this should be adjusted to use the new serializer in
+  `ocean.util.serialize`. (The stream and plugin serialization facilities of
+  `StructSerializer` remain unchanged.)
+
 New Features
 ============